requirements.txt

langchain== 0.3.18
Flask_Cors==5.0.0
python-dotenv==1.0.1
openai>=1.35.14,<2.0.0
gunicorn==23.0.0
pinecone-client==5.0.1
langchain==0.3.18
boto3==1.36.19
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
protobuf
grpcio
grpcio-tools
protoc-gen-openapiv2
langchain-openai==0.3.5
